requirements.txt
redis
cassandra-driver
statsd
tweepy
celery[redis]
requests
